+++ src/pages/SignUpPage/hooks/index.ts
@@ -3,3 +3,4 @@ export { default as useFullNameCheck } from './useFullNameCheck';
 export { default as usePasswordCheck } from './usePasswordCheck';
 export { default as useClickEye } from './useClickEye';
 export { default as useAllValidationPass } from './useAllValidationPass';
+export { default as useSubmit } from './useSubmit';
